backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==23.0.0
whitenoise==6.11.0
Pillow==12.0.0
dj-database-url==3.0.1
psycopg[binary]==3.2.12
